pyproject.toml
[project]
name = "nonebot-plugin-cd-managaer"
version = "0.1.0"
description = "通用cd和撤回管理"
authors = [
    {name = "Cvandia", email = "106718176+Cvandia@users.noreply.github.com"},
]
dependencies = ["nonebot2>=2.2.0", "nonebot-plugin-alconna>=0.53.1"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[dependency-groups]
dev = ["nonebot2[fastapi]>=2.4.0", "nb-cli>=1.4.2"]
